Enforce restriction on strings in all builds
I think we should check this in all builds, since otherwise they won't
work. (We *could* use the ORG command instead of the ERROR command for
non-SD builds, since in other builds we have space to waste. But for now
let's see if we can avoid wasting space in any build.)