Fix bug (dp11 post Tue Mar 01, 2016 9:57 am)
The earlier rework of the command table handling to use JMP (abs, X)
failed to take into account the fact that the first byte following the
command name was expected to have the top bit set. I also managed to
mess up the final entry in command_table as well.

I've tested this in BeebEm with a SCSI hard drive (briefly).